refactor pt 14: split device host loop into device_loop.py
- move run_window, CFL host control (_cfl_maxima*, _host_dt), _check_supported, _eval_forcing and the window compile cache (device_kernels.py 697-921, verbatim) into device_loop.py
- device_kernels keeps the traced substeps + make_step (~690 lines)
- device_step shim now re-exports run_window from device_loop; __all__ unchanged
- gates: device<->device bit-identity (fast set, pre-split baseline), 11/11 device equivalence + fullrun tests, numpy fast bit-diff
