+++ changelog.d/device_loop_split.changed.md
@@ -1,0 +1,3 @@
+Split the jax-device host window driver (`run_window`, CFL/dt host control,
+support guard, forcing eval, compile cache) out of `device_kernels.py` into
+`backends/jax_ops/device_loop.py`; `device_step` re-exports are unchanged.
